Transformer: Merge ImportResolver with IdentifierTypeResolver
Handling module import when indices can't be resolved caused (probably) unsolvable issues when a compiled modules was imported into another one

All code generation test passes
