--- include/NZSL/Ast/Transformations/ImportResolverTransformer.inl
@@ -1,16 +1,0 @@
-// Copyright (C) 2025 Jérôme "SirLynix" Leclercq (lynix680@gmail.com)
-// This file is part of the "Nazara Shading Language" project
-// For conditions of distribution and use, see copyright notice in Config.hpp
-
-namespace nzsl::Ast
-{
-	inline ImportResolverTransformer::ImportResolverTransformer() :
-	Transformer(true)
-	{
-	}
-
-	inline bool ImportResolverTransformer::Transform(Module& module, Context& context, std::string* error)
-	{
-		return Transform(module, context, {}, error);
-	}
-}
