+++ tests/src/Tests/ShaderUtils.hpp
@@ -9,7 +9,6 @@
 #include <NZSL/Ast/Transformations/BindingResolverTransformer.hpp>
 #include <NZSL/Ast/Transformations/ConstantRemovalTransformer.hpp>
 #include <NZSL/Ast/Transformations/IdentifierTypeResolverTransformer.hpp>
-#include <NZSL/Ast/Transformations/ImportResolverTransformer.hpp>
 #include <spirv-tools/libspirv.hpp>
 #include <filesystem>
 
@@ -23,7 +22,6 @@ struct ResolveOptions
 	const nzsl::Ast::BindingResolverTransformer::Options* bindingResolverOptions = &defaultBindingResolverOptions;
 	const nzsl::Ast::ConstantRemovalTransformer::Options* constantRemovalOptions = nullptr;
 	const nzsl::Ast::IdentifierTypeResolverTransformer::Options* identifierResolverOptions = &defaultIdentifierResolveOptions;
-	const nzsl::Ast::ImportResolverTransformer::Options* importOptions = nullptr;
 };
 
 void ExpectGLSL(nzsl::ShaderStageType stageType, nzsl::Ast::Module& shader, std::string_view expectedOutput, const nzsl::ShaderWriter::States& options = {}, const nzsl::GlslWriter::Environment& env = {}, bool testShaderCompilation = true);
